requirements.txt

flask
torch
realesrgan
Pillow
